do not delete working sub directory in cleanup method

+++ src/main/java/fr/ans/psc/pscextract/service/utils/FileNamesUtil.java
@@ -56,7 +56,9 @@ public class FileNamesUtil {
         }
 
         for(File file : listOfFiles) {
-            file.delete();
+            if (!file.isDirectory()) {
+                file.delete();
+            }
         }
     }
